Put the usage method to work, added better output.
The usage method is now called when the program is launched with too
many or too few arguments. Additionally, there is more output when the
program is saving images, so that the program does not give the
impression that it has stalled. This output includes the number of
images saved, the file name, and the description from the tweet the
image was saved from.